Refactor resource constraint extensions per tfoote feedback
- Move ShmSize, CpuLimits, and MemoryLimits to separate constraint_extensions.py file
- Create reusable validate_memory_format() helper function to avoid code duplication
- Enhance memory validation to support Docker byte-value formats (2b, 1024kb, 300m, 1gb)
- Update CPU error message to "Expected a floating point number (e.g., '1.5')"
- Update setup.py entry points to reference new constraint_extensions module

Addresses all feedback from tfoote in PR #331 review comments.
