+++ setup.py
@@ -45,7 +45,7 @@ kwargs = {
             'detect_docker_image_os = rocker.cli:detect_image_os',
         ],
         'rocker.extensions': [
-            'cpu_limits = rocker.extensions:CpuLimits',
+            'cpu_limits = rocker.constraint_extensions:CpuLimits',
             'cuda = rocker.nvidia_extension:Cuda',
             'detach = rocker.extensions:Detach',
             'devices = rocker.extensions:Devices',
@@ -57,7 +57,7 @@ kwargs = {
             'home = rocker.extensions:HomeDir',
             'hostname = rocker.extensions:Hostname',
             'ipc = rocker.extensions:Ipc',
-            'memory_limits = rocker.extensions:MemoryLimits',
+            'memory_limits = rocker.constraint_extensions:MemoryLimits',
             'name = rocker.extensions:Name',
             'network = rocker.extensions:Network',
             'nvidia = rocker.nvidia_extension:Nvidia',
@@ -65,7 +65,7 @@ kwargs = {
             'privileged = rocker.extensions:Privileged',
             'pulse = rocker.extensions:PulseAudio',
             'rmw = rocker.rmw_extension:RMW',
-            'shm_size = rocker.extensions:ShmSize',
+            'shm_size = rocker.constraint_extensions:ShmSize',
             'ssh = rocker.ssh_extension:Ssh',
             'ulimit = rocker.ulimit_extension:Ulimit',
             'user = rocker.extensions:User',
